refactor: delete runtime-dead src/tools/ssl-analysis.ts duplicate
src/tools/ssl-analysis.ts duplicated getHttpRedirectFindings,
getHttpsFindings, and getHttpsErrorFinding from
packages/dns-checks/src/checks/ssl-analysis.ts. Its only importer was
test/ssl-analysis.spec.ts — check-ssl.ts has always delegated entirely
to the package's checkSSL, so no runtime path touched the dead copy.

Follow-up to #806/#819: that fix patched the no-content (204/205)
guard into the package copy of getHttpRedirectFindings only, since
that's the one actually reachable at runtime. Leaving the dead copy in
place meant it kept diverging from the live implementation, so a
future importer would have resurrected the exact false-finding defect
#806/#819 just fixed.

Same precedent as #621, which deleted the analogous dead
src/tools/mta-sts-analysis.ts duplicate outright rather than
re-exporting from the package.

getHttpsFindings and getHttpsErrorFinding had no direct unit test
anywhere in the package (only indirect exercise via checkSSL
integration tests and the worker-level test/check-ssl.spec.ts, which
is unaffected by this change and stays). Ported the deleted spec's
unique assertions for both into a new
packages/dns-checks/src/__tests__/checks/ssl-analysis.test.ts.
getHttpRedirectFindings needed no porting: its behavior is already
covered directly by ssl-no-content.test.ts (#819).
